backing up
git-svn-id: https://svn.code.sf.net/p/okai/code/n02@34 0ac42f08-cc4f-0410-a013-8b35e0a3db03

+++ src/modules/modDefStress/juceModDefStressTest.cpp
@@ -253,7 +253,7 @@ void juceModDefStressTest::sliderValueChanged (Slider* sliderThatWasMoved)
     if (sliderThatWasMoved == sldDelay)
     {
         //[UserSliderCode_sldDelay] -- add your slider handling code here..
-		n02::defStressT::delay = sldDelay->getValue();
+		n02::defStressT::delay = (int)sldDelay->getValue();
         //[/UserSliderCode_sldDelay]
     }
